refactor: Change the way of passport authenticate
Add passportConfig file for strategy log in

Add authenticate middleware and remove it from login route
