+++ TODO.md
@@ -1,0 +1,22 @@
+Refactor to react from ejs
+
+    [] refactor: login route
+    [] refactor: login frontend
+
+    [] refactor: register route
+    [] refactor: register frontend
+
+    [] refactor: chat rooms route
+    [] refactor: chat rooms frontend
+
+    [] refactor: index(chats) route
+    [] refactor: index(chats) frontend
+
+    [] refactor: profil route
+    [] refactor: profil frontend
+
+    [] refactor: ranges route
+    [] refactor: ranges frontend
+
+    [] refactor: users route
+    [] refactor: users frontend
